Refactor reading TIFF metadata
