+++ src/ifd.rs
@@ -1,5 +1,4 @@
 use std::collections::HashMap;
-use std::io::Read;
 use std::ops::Range;
 
 use bytes::Bytes;
@@ -7,52 +6,15 @@ use num_enum::TryFromPrimitive;
 
 use crate::error::{AsyncTiffError, AsyncTiffResult};
 use crate::geo::{GeoKeyDirectory, GeoKeyTag};
-use crate::reader::{AsyncCursor, AsyncFileReader};
+use crate::reader::AsyncFileReader;
 use crate::tiff::tags::{
     CompressionMethod, PhotometricInterpretation, PlanarConfiguration, Predictor, ResolutionUnit,
-    SampleFormat, Tag, Type,
+    SampleFormat, Tag,
 };
 use crate::tiff::{TiffError, Value};
 use crate::tile::Tile;
 
 const DOCUMENT_NAME: u16 = 269;
-
-/// A collection of all the IFD
-// TODO: maybe separate out the primary/first image IFD out of the vec, as that one should have
-// geospatial metadata?
-#[derive(Debug, Clone)]
-pub struct ImageFileDirectories {
-    /// There's always at least one IFD in a TIFF. We store this separately
-    ifds: Vec<ImageFileDirectory>,
-    // Is it guaranteed that if masks exist that there will be one per image IFD? Or could there be
-    // different numbers of image ifds and mask ifds?
-    // mask_ifds: Option<Vec<IFD>>,
-}
-
-impl AsRef<[ImageFileDirectory]> for ImageFileDirectories {
-    fn as_ref(&self) -> &[ImageFileDirectory] {
-        &self.ifds
-    }
-}
-
-impl ImageFileDirectories {
-    pub(crate) async fn open(
-        cursor: &mut AsyncCursor,
-        ifd_offset: u64,
-        bigtiff: bool,
-    ) -> AsyncTiffResult<Self> {
-        let mut next_ifd_offset = Some(ifd_offset);
-
-        let mut ifds = vec![];
-        while let Some(offset) = next_ifd_offset {
-            let ifd = ImageFileDirectory::read(cursor, offset, bigtiff).await?;
-            next_ifd_offset = ifd.next_ifd_offset();
-            ifds.push(ifd);
-        }
-
-        Ok(Self { ifds })
-    }
-}
 
 /// An ImageFileDirectory representing Image content
 // The ordering of these tags matches the sorted order in TIFF spec Appendix A
@@ -177,69 +139,11 @@ pub struct ImageFileDirectory {
     // no_data
     // gdal_metadata
     pub(crate) other_tags: HashMap<Tag, Value>,
-
-    pub(crate) next_ifd_offset: Option<u64>,
 }
 
 impl ImageFileDirectory {
-    /// Read and parse the IFD starting at the given file offset
-    async fn read(
-        cursor: &mut AsyncCursor,
-        ifd_start: u64,
-        bigtiff: bool,
-    ) -> AsyncTiffResult<Self> {
-        cursor.seek(ifd_start);
-
-        let tag_count = if bigtiff {
-            cursor.read_u64().await?
-        } else {
-            cursor.read_u16().await?.into()
-        };
-        let mut tags = HashMap::with_capacity(tag_count as usize);
-        for _ in 0..tag_count {
-            let (tag_name, tag_value) = read_tag(cursor, bigtiff).await?;
-            tags.insert(tag_name, tag_value);
-        }
-
-        // Tag   2 bytes
-        // Type  2 bytes
-        // Count:
-        //  - bigtiff: 8 bytes
-        //  - else: 4 bytes
-        // Value:
-        //  - bigtiff: 8 bytes either a pointer the value itself
-        //  - else: 4 bytes either a pointer the value itself
-        let ifd_entry_byte_size = if bigtiff { 20 } else { 12 };
-        // The size of `tag_count` that we read above
-        let tag_count_byte_size = if bigtiff { 8 } else { 2 };
-
-        // Reset the cursor position before reading the next ifd offset
-        cursor.seek(ifd_start + (ifd_entry_byte_size * tag_count) + tag_count_byte_size);
-
-        let next_ifd_offset = if bigtiff {
-            cursor.read_u64().await?
-        } else {
-            cursor.read_u32().await?.into()
-        };
-
-        // If the ifd_offset is 0, stop
-        let next_ifd_offset = if next_ifd_offset == 0 {
-            None
-        } else {
-            Some(next_ifd_offset)
-        };
-
-        Self::from_tags(tags, next_ifd_offset)
-    }
-
-    fn next_ifd_offset(&self) -> Option<u64> {
-        self.next_ifd_offset
-    }
-
-    fn from_tags(
-        mut tag_data: HashMap<Tag, Value>,
-        next_ifd_offset: Option<u64>,
-    ) -> AsyncTiffResult<Self> {
+    /// Create a new ImageFileDirectory from tag data
+    pub fn from_tags(tag_data: HashMap<Tag, Value>) -> AsyncTiffResult<Self> {
         let mut new_subfile_type = None;
         let mut image_width = None;
         let mut image_height = None;
@@ -281,7 +185,10 @@ impl ImageFileDirectory {
 
         let mut other_tags = HashMap::new();
 
-        tag_data.drain().try_for_each(|(tag, value)| {
+        // for x in tag_data.into_iter() {
+
+        // }
+        tag_data.into_iter().try_for_each(|(tag, value)| {
             match tag {
                 Tag::NewSubfileType => new_subfile_type = Some(value.into_u32()?),
                 Tag::ImageWidth => image_width = Some(value.into_u32()?),
@@ -485,7 +392,6 @@ impl ImageFileDirectory {
             model_pixel_scale,
             model_tiepoint,
             other_tags,
-            next_ifd_offset,
         })
     }
 
@@ -779,7 +685,7 @@ impl ImageFileDirectory {
         let range = self
             .get_tile_byte_range(x, y)
             .ok_or(AsyncTiffError::General("Not a tiled TIFF".to_string()))?;
-        let compressed_bytes = reader.get_image_bytes(range).await?;
+        let compressed_bytes = reader.get_bytes(range).await?;
         Ok(Tile {
             x,
             y,
@@ -809,8 +715,8 @@ impl ImageFileDirectory {
             })
             .collect::<AsyncTiffResult<Vec<_>>>()?;
 
-        // 2: Fetch using `get_image_byte_ranges`
-        let buffers = reader.get_image_byte_ranges(byte_ranges).await?;
+        // 2: Fetch using `get_byte_ranges`
+        let buffers = reader.get_byte_ranges(byte_ranges).await?;
 
         // 3: Create tile objects
         let mut tiles = vec![];
@@ -834,372 +740,5 @@ impl ImageFileDirectory {
         let x_count = (self.image_width as f64 / self.tile_width? as f64).ceil();
         let y_count = (self.image_height as f64 / self.tile_height? as f64).ceil();
         Some((x_count as usize, y_count as usize))
-    }
-}
-
-/// Read a single tag from the cursor
-async fn read_tag(cursor: &mut AsyncCursor, bigtiff: bool) -> AsyncTiffResult<(Tag, Value)> {
-    let start_cursor_position = cursor.position();
-
-    let tag_name = Tag::from_u16_exhaustive(cursor.read_u16().await?);
-
-    let tag_type_code = cursor.read_u16().await?;
-    let tag_type = Type::from_u16(tag_type_code).expect(
-        "Unknown tag type {tag_type_code}. TODO: we should skip entries with unknown tag types.",
-    );
-    let count = if bigtiff {
-        cursor.read_u64().await?
-    } else {
-        cursor.read_u32().await?.into()
-    };
-
-    let tag_value = read_tag_value(cursor, tag_type, count, bigtiff).await?;
-
-    // TODO: better handle management of cursor state
-    let ifd_entry_size = if bigtiff { 20 } else { 12 };
-    cursor.seek(start_cursor_position + ifd_entry_size);
-
-    Ok((tag_name, tag_value))
-}
-
-/// Read a tag's value from the cursor
-///
-/// NOTE: this does not maintain cursor state
-// This is derived from the upstream tiff crate:
-// https://github.com/image-rs/image-tiff/blob/6dc7a266d30291db1e706c8133357931f9e2a053/src/decoder/ifd.rs#L369-L639
-async fn read_tag_value(
-    cursor: &mut AsyncCursor,
-    tag_type: Type,
-    count: u64,
-    bigtiff: bool,
-) -> AsyncTiffResult<Value> {
-    // Case 1: there are no values so we can return immediately.
-    if count == 0 {
-        return Ok(Value::List(vec![]));
-    }
-
-    let tag_size = match tag_type {
-        Type::BYTE | Type::SBYTE | Type::ASCII | Type::UNDEFINED => 1,
-        Type::SHORT | Type::SSHORT => 2,
-        Type::LONG | Type::SLONG | Type::FLOAT | Type::IFD => 4,
-        Type::LONG8
-        | Type::SLONG8
-        | Type::DOUBLE
-        | Type::RATIONAL
-        | Type::SRATIONAL
-        | Type::IFD8 => 8,
-    };
-
-    let value_byte_length = count.checked_mul(tag_size).unwrap();
-
-    // Case 2: there is one value.
-    if count == 1 {
-        // 2a: the value is 5-8 bytes and we're in BigTiff mode.
-        if bigtiff && value_byte_length > 4 && value_byte_length <= 8 {
-            let mut data = cursor.read(value_byte_length).await?;
-
-            return Ok(match tag_type {
-                Type::LONG8 => Value::UnsignedBig(data.read_u64()?),
-                Type::SLONG8 => Value::SignedBig(data.read_i64()?),
-                Type::DOUBLE => Value::Double(data.read_f64()?),
-                Type::RATIONAL => Value::Rational(data.read_u32()?, data.read_u32()?),
-                Type::SRATIONAL => Value::SRational(data.read_i32()?, data.read_i32()?),
-                Type::IFD8 => Value::IfdBig(data.read_u64()?),
-                Type::BYTE
-                | Type::SBYTE
-                | Type::ASCII
-                | Type::UNDEFINED
-                | Type::SHORT
-                | Type::SSHORT
-                | Type::LONG
-                | Type::SLONG
-                | Type::FLOAT
-                | Type::IFD => unreachable!(),
-            });
-        }
-
-        // NOTE: we should only be reading value_byte_length when it's 4 bytes or fewer. Right now
-        // we're reading even if it's 8 bytes, but then only using the first 4 bytes of this
-        // buffer.
-        let mut data = cursor.read(value_byte_length).await?;
-
-        // 2b: the value is at most 4 bytes or doesn't fit in the offset field.
-        return Ok(match tag_type {
-            Type::BYTE | Type::UNDEFINED => Value::Byte(data.read_u8()?),
-            Type::SBYTE => Value::Signed(data.read_i8()? as i32),
-            Type::SHORT => Value::Short(data.read_u16()?),
-            Type::SSHORT => Value::Signed(data.read_i16()? as i32),
-            Type::LONG => Value::Unsigned(data.read_u32()?),
-            Type::SLONG => Value::Signed(data.read_i32()?),
-            Type::FLOAT => Value::Float(data.read_f32()?),
-            Type::ASCII => {
-                if data.as_ref()[0] == 0 {
-                    Value::Ascii("".to_string())
-                } else {
-                    panic!("Invalid tag");
-                    // return Err(TiffError::FormatError(TiffFormatError::InvalidTag));
-                }
-            }
-            Type::LONG8 => {
-                let offset = data.read_u32()?;
-                cursor.seek(offset as _);
-                Value::UnsignedBig(cursor.read_u64().await?)
-            }
-            Type::SLONG8 => {
-                let offset = data.read_u32()?;
-                cursor.seek(offset as _);
-                Value::SignedBig(cursor.read_i64().await?)
-            }
-            Type::DOUBLE => {
-                let offset = data.read_u32()?;
-                cursor.seek(offset as _);
-                Value::Double(cursor.read_f64().await?)
-            }
-            Type::RATIONAL => {
-                let offset = data.read_u32()?;
-                cursor.seek(offset as _);
-                let numerator = cursor.read_u32().await?;
-                let denominator = cursor.read_u32().await?;
-                Value::Rational(numerator, denominator)
-            }
-            Type::SRATIONAL => {
-                let offset = data.read_u32()?;
-                cursor.seek(offset as _);
-                let numerator = cursor.read_i32().await?;
-                let denominator = cursor.read_i32().await?;
-                Value::SRational(numerator, denominator)
-            }
-            Type::IFD => Value::Ifd(data.read_u32()?),
-            Type::IFD8 => {
-                let offset = data.read_u32()?;
-                cursor.seek(offset as _);
-                Value::IfdBig(cursor.read_u64().await?)
-            }
-        });
-    }
-
-    // Case 3: There is more than one value, but it fits in the offset field.
-    if value_byte_length <= 4 || bigtiff && value_byte_length <= 8 {
-        let mut data = cursor.read(value_byte_length).await?;
-        if bigtiff {
-            cursor.advance(8 - value_byte_length);
-        } else {
-            cursor.advance(4 - value_byte_length);
-        }
-
-        match tag_type {
-            Type::BYTE | Type::UNDEFINED => {
-                return {
-                    Ok(Value::List(
-                        (0..count)
-                            .map(|_| Value::Byte(data.read_u8().unwrap()))
-                            .collect(),
-                    ))
-                };
-            }
-            Type::SBYTE => {
-                return {
-                    Ok(Value::List(
-                        (0..count)
-                            .map(|_| Value::Signed(data.read_i8().unwrap() as i32))
-                            .collect(),
-                    ))
-                }
-            }
-            Type::ASCII => {
-                let mut buf = vec![0; count as usize];
-                data.read_exact(&mut buf)?;
-                if buf.is_ascii() && buf.ends_with(&[0]) {
-                    let v = std::str::from_utf8(&buf)
-                        .map_err(|err| AsyncTiffError::General(err.to_string()))?;
-                    let v = v.trim_matches(char::from(0));
-                    return Ok(Value::Ascii(v.into()));
-                } else {
-                    panic!("Invalid tag");
-                    // return Err(TiffError::FormatError(TiffFormatError::InvalidTag));
-                }
-            }
-            Type::SHORT => {
-                let mut v = Vec::new();
-                for _ in 0..count {
-                    v.push(Value::Short(data.read_u16()?));
-                }
-                return Ok(Value::List(v));
-            }
-            Type::SSHORT => {
-                let mut v = Vec::new();
-                for _ in 0..count {
-                    v.push(Value::Signed(i32::from(data.read_i16()?)));
-                }
-                return Ok(Value::List(v));
-            }
-            Type::LONG => {
-                let mut v = Vec::new();
-                for _ in 0..count {
-                    v.push(Value::Unsigned(data.read_u32()?));
-                }
-                return Ok(Value::List(v));
-            }
-            Type::SLONG => {
-                let mut v = Vec::new();
-                for _ in 0..count {
-                    v.push(Value::Signed(data.read_i32()?));
-                }
-                return Ok(Value::List(v));
-            }
-            Type::FLOAT => {
-                let mut v = Vec::new();
-                for _ in 0..count {
-                    v.push(Value::Float(data.read_f32()?));
-                }
-                return Ok(Value::List(v));
-            }
-            Type::IFD => {
-                let mut v = Vec::new();
-                for _ in 0..count {
-                    v.push(Value::Ifd(data.read_u32()?));
-                }
-                return Ok(Value::List(v));
-            }
-            Type::LONG8
-            | Type::SLONG8
-            | Type::RATIONAL
-            | Type::SRATIONAL
-            | Type::DOUBLE
-            | Type::IFD8 => {
-                unreachable!()
-            }
-        }
-    }
-
-    // Seek cursor
-    let offset = if bigtiff {
-        cursor.read_u64().await?
-    } else {
-        cursor.read_u32().await?.into()
-    };
-    cursor.seek(offset);
-
-    // Case 4: there is more than one value, and it doesn't fit in the offset field.
-    match tag_type {
-        // TODO check if this could give wrong results
-        // at a different endianess of file/computer.
-        Type::BYTE | Type::UNDEFINED => {
-            let mut v = Vec::with_capacity(count as _);
-            for _ in 0..count {
-                v.push(Value::Byte(cursor.read_u8().await?))
-            }
-            Ok(Value::List(v))
-        }
-        Type::SBYTE => {
-            let mut v = Vec::with_capacity(count as _);
-            for _ in 0..count {
-                v.push(Value::Signed(cursor.read_i8().await? as i32))
-            }
-            Ok(Value::List(v))
-        }
-        Type::SHORT => {
-            let mut v = Vec::with_capacity(count as _);
-            for _ in 0..count {
-                v.push(Value::Short(cursor.read_u16().await?))
-            }
-            Ok(Value::List(v))
-        }
-        Type::SSHORT => {
-            let mut v = Vec::with_capacity(count as _);
-            for _ in 0..count {
-                v.push(Value::Signed(cursor.read_i16().await? as i32))
-            }
-            Ok(Value::List(v))
-        }
-        Type::LONG => {
-            let mut v = Vec::with_capacity(count as _);
-            for _ in 0..count {
-                v.push(Value::Unsigned(cursor.read_u32().await?))
-            }
-            Ok(Value::List(v))
-        }
-        Type::SLONG => {
-            let mut v = Vec::with_capacity(count as _);
-            for _ in 0..count {
-                v.push(Value::Signed(cursor.read_i32().await?))
-            }
-            Ok(Value::List(v))
-        }
-        Type::FLOAT => {
-            let mut v = Vec::with_capacity(count as _);
-            for _ in 0..count {
-                v.push(Value::Float(cursor.read_f32().await?))
-            }
-            Ok(Value::List(v))
-        }
-        Type::DOUBLE => {
-            let mut v = Vec::with_capacity(count as _);
-            for _ in 0..count {
-                v.push(Value::Double(cursor.read_f64().await?))
-            }
-            Ok(Value::List(v))
-        }
-        Type::RATIONAL => {
-            let mut v = Vec::with_capacity(count as _);
-            for _ in 0..count {
-                v.push(Value::Rational(
-                    cursor.read_u32().await?,
-                    cursor.read_u32().await?,
-                ))
-            }
-            Ok(Value::List(v))
-        }
-        Type::SRATIONAL => {
-            let mut v = Vec::with_capacity(count as _);
-            for _ in 0..count {
-                v.push(Value::SRational(
-                    cursor.read_i32().await?,
-                    cursor.read_i32().await?,
-                ))
-            }
-            Ok(Value::List(v))
-        }
-        Type::LONG8 => {
-            let mut v = Vec::with_capacity(count as _);
-            for _ in 0..count {
-                v.push(Value::UnsignedBig(cursor.read_u64().await?))
-            }
-            Ok(Value::List(v))
-        }
-        Type::SLONG8 => {
-            let mut v = Vec::with_capacity(count as _);
-            for _ in 0..count {
-                v.push(Value::SignedBig(cursor.read_i64().await?))
-            }
-            Ok(Value::List(v))
-        }
-        Type::IFD => {
-            let mut v = Vec::with_capacity(count as _);
-            for _ in 0..count {
-                v.push(Value::Ifd(cursor.read_u32().await?))
-            }
-            Ok(Value::List(v))
-        }
-        Type::IFD8 => {
-            let mut v = Vec::with_capacity(count as _);
-            for _ in 0..count {
-                v.push(Value::IfdBig(cursor.read_u64().await?))
-            }
-            Ok(Value::List(v))
-        }
-        Type::ASCII => {
-            let mut out = vec![0; count as _];
-            let mut buf = cursor.read(count).await?;
-            buf.read_exact(&mut out)?;
-
-            // Strings may be null-terminated, so we trim anything downstream of the null byte
-            if let Some(first) = out.iter().position(|&b| b == 0) {
-                out.truncate(first);
-            }
-            Ok(Value::Ascii(
-                String::from_utf8(out).map_err(|err| AsyncTiffError::General(err.to_string()))?,
-            ))
-        }
     }
 }
