+++ src/metadata/mod.rs
@@ -1,0 +1,7 @@
+//! TIFF metadata API
+
+mod fetch;
+mod reader;
+
+pub use fetch::{MetadataFetch, PrefetchMetadataFetch};
+pub use reader::{ImageFileDirectoryReader, TiffMetadataReader};
